Moves functionality to composable
