+++ ppr-ui/src/composables/mhrInformation/useMhrUnitNote.ts
@@ -1,5 +1,5 @@
 import { UnitNoteDocTypes, UnitNoteStatusTypes } from '@/enums/unitNoteDocTypes'
-import { PartyIF, UnitNoteIF, UnitNoteRegistrationIF } from '@/interfaces'
+import { GroupedNotesIF, PartyIF, UnitNoteIF, UnitNoteRegistrationIF } from '@/interfaces'
 import { useStore } from '@/store/store'
 import { deleteEmptyProperties, submitMhrUnitNote } from '@/utils'
 import { storeToRefs } from 'pinia'
@@ -116,9 +116,46 @@ export const useMhrUnitNote = () => {
       UnitNoteDocTypes.EXTENSION_TO_NOTICE_OF_CAUTION].includes(note.documentType)
   }
 
+  // Groups unit notes for the panels
+  const groupUnitNotes = (unitNotes: UnitNoteIF[]): GroupedNotesIF[] => {
+    // The notes should already be in order by creation date and time (filter preserves order)
+    const noticeOfCautionNotes = unitNotes.filter((note) => isNoticeOfCautionOrRelatedDocType(note))
+
+    let primaryUnitNote: UnitNoteIF = null
+    let additionalUnitNotes: UnitNoteIF[] = []
+
+    const groupedNoticeOfCautions: GroupedNotesIF[] = []
+
+    // NOTICE_OF_CAUTION is used as an interval for grouping the notes
+    // When a NOTICE_OF_CAUTION is encountered, the group is added
+    // and a new group is started on the next iteration
+    noticeOfCautionNotes.forEach((note) => {
+      if (!primaryUnitNote) { primaryUnitNote = note } else { additionalUnitNotes.push(note) }
+
+      if (note.documentType === UnitNoteDocTypes.NOTICE_OF_CAUTION) {
+        groupedNoticeOfCautions.push({ primaryUnitNote, additionalUnitNotes })
+        primaryUnitNote = null
+        additionalUnitNotes = []
+      }
+    })
+
+    const nonNoticeOfCautionUnitNotes = unitNotes.filter((note) => !isNoticeOfCautionOrRelatedDocType(note))
+
+    const groupedUnitNotes: GroupedNotesIF[] = nonNoticeOfCautionUnitNotes.map((note) => {
+      return { primaryUnitNote: note }
+    })
+
+    // Adds the notice of caution notes to the other unit notes and sort in descending creation time
+    return groupedUnitNotes.concat(groupedNoticeOfCautions).sort((note1, note2) =>
+      new Date(note2.primaryUnitNote.createDateTime).getTime() -
+       new Date(note1.primaryUnitNote.createDateTime).getTime()
+    )
+  }
+
   return {
     initUnitNote,
     buildApiDataAndSubmit,
+    groupUnitNotes,
     isNoticeOfCautionOrRelatedDocType
   }
 }
